Tested game on Mac:
Changed the turn speed of the beacon from 180 to 155 to increase shooting accuracy.

Fixed a bug where only 1 light beacon would be activated at a time. Now multiple might beacons can be activated. Light beacons have a bool variable called "activated". It was previously public static. But since no one needed to know that, I changed it to private. That allows multiple beacons to be activated.